Firmware_Analyser/firmwareanalyzer/requirements.txt
asgiref==3.7.2
boto3==1.28.48
botocore==1.31.48
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
Deprecated==1.2.14
diff-match-patch==20230430
Django==4.2.4
django-cors-headers==3.13.0
django-environ==0.10.0
django-filter==23.1
django-import-export==3.1.0
django-oauth-toolkit==2.3.0
django-ses==3.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-social-oauth2==2.1.3
ecdsa==0.18.0
et-xmlfile==1.1.0
idna==3.4
jmespath==1.0.1
jwcrypto==1.5.0
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupPy==1.14
mdurl==0.1.2
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
Pillow==9.4.0
pyasn1==0.5.0
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.3
rsa==4.9
s3transfer==0.6.2
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
tablib==3.5.0
tzdata==2023.3
urllib3==1.26.16
wrapt==1.15.0
xlrd==2.0.1
xlwt==1.3.0
